+++ src/MasterChef.Domain/Models/RecipeModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using MasterChef.Domain.Entities;
