+++ src/MasterChef.Dto/RecipeDto.cs
@@ -12,18 +12,3 @@ public class RecipeDto
 
     public UserDto? User { get; set; }
 }
-
-public class UserDto
-{
-    public string? Username { get; set; }
-    public string? ExternalId { get; set; }
-}
-public class IngredientDto
-{
-    public int? Id { get; set; }
-    public string? Name { get; set; }
-    public string? Description { get; set; }
-    public decimal? Weight { get; set; }
-    public int? Quantity { get; set; }
-    public int? RecipeId { get; set; }
-}
