+++ src/MasterChef.Dto/UserDto.cs
@@ -1,0 +1,7 @@
+namespace MasterChef.Dto;
+
+public class UserDto
+{
+    public string? Username { get; set; }
+    public string? ExternalId { get; set; }
+}
